Abstract promoteDraftsComponent
