+++ src/use-main-actionmenu.js
@@ -1,5 +1,4 @@
-const { SyntaxKind } = require('ts-morph');
-const updateImportDeclaration = require('./utils/change-drafts-import-to-main');
+const promoteDraftsComponent = require('./utils/promote-drafts-component');
 
 const componentImportNames = [
   'ActionMenu',
@@ -11,20 +10,7 @@ const componentImportNames = [
 const draftsFileName = 'ActionMenu2';
 
 const transform = (project) => {
-  const sourceFiles = project.getSourceFiles();
-
-  sourceFiles.forEach((sourceFile) => {
-    try {
-      sourceFile.getDescendantsOfKind(SyntaxKind.ImportDeclaration).forEach((declaration) => {
-        declaration = updateImportDeclaration(declaration, sourceFile, componentImportNames, draftsFileName);
-      });
-
-      // save source back to file
-      sourceFile.saveSync();
-    } catch (error) {
-      console.log(error);
-    }
-  });
+  declaration = promoteDraftsComponent(project, componentImportNames, draftsFileName);
 };
 
 module.exports = transform;
