+++ src/use-main-button.js
@@ -1,24 +1,10 @@
-const { SyntaxKind } = require('ts-morph');
-const updateImportDeclaration = require('./utils/change-drafts-import-to-main');
+const promoteDraftsComponent = require('./utils/promote-drafts-component');
 
 const componentImportNames = ['Button', 'ButtonProps', 'IconButton', 'IconButtonProps', 'LinkButton'];
 const draftsFileName = 'Button2';
 
 const transform = (project) => {
-  const sourceFiles = project.getSourceFiles();
-
-  sourceFiles.forEach((sourceFile) => {
-    try {
-      sourceFile.getDescendantsOfKind(SyntaxKind.ImportDeclaration).forEach((declaration) => {
-        declaration = updateImportDeclaration(declaration, sourceFile, componentImportNames, draftsFileName);
-      });
-
-      // save source back to file
-      sourceFile.saveSync();
-    } catch (error) {
-      console.log(error);
-    }
-  });
+  declaration = promoteDraftsComponent(project, componentImportNames, draftsFileName);
 };
 
 module.exports = transform;
