+++ src/use-main-pagelayout.js
@@ -1,5 +1,4 @@
-const { SyntaxKind } = require('ts-morph');
-const updateImportDeclaration = require('./utils/change-drafts-import-to-main');
+const promoteDraftsComponent = require('./utils/promote-drafts-component');
 
 const componentImportNames = [
   'PageLayout',
@@ -12,20 +11,7 @@ const componentImportNames = [
 const draftsFileName = 'PageLayout';
 
 const transform = (project) => {
-  const sourceFiles = project.getSourceFiles();
-
-  sourceFiles.forEach((sourceFile) => {
-    try {
-      sourceFile.getDescendantsOfKind(SyntaxKind.ImportDeclaration).forEach((declaration) => {
-        declaration = updateImportDeclaration(declaration, sourceFile, componentImportNames, draftsFileName);
-      });
-
-      // save source back to file
-      sourceFile.saveSync();
-    } catch (error) {
-      console.log(error);
-    }
-  });
+  declaration = promoteDraftsComponent(project, componentImportNames, draftsFileName);
 };
 
 module.exports = transform;
